fixed error which occurs when using Flexbox nested in styled(Flexbox) with SSR:
following Error was raised without this fix:
Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.